final/commit5

+++ project_movies/src/main/java/com/example/project_movies/domain/CommentEntity.java
@@ -30,7 +30,7 @@ public class CommentEntity {
     @JoinColumn(name = "movie_id", nullable = false)
     private MovieEntity movie;
 
-    public CommentEntity(String comment, Integer rating, MovieEntity movie) {
+    public CommentEntity(String comment, double rating, MovieEntity movie) {
         this.comment = comment;
         this.rating = rating;
         this.movie = movie;
